# [alpaka] Implement "external work" producers and asynchronous products