--- a/src/alpaka/AlpakaCore/EventCache.h
+++ b/src/alpaka/AlpakaCore/EventCache.h
@@ -54,7 +54,7 @@ namespace cms::alpakatools {
 
   private:
     std::shared_ptr<Event> makeOrGet(Device dev) {
-      return cache_[cms::alpakatools::getDevIndex(dev)].makeOrGet([dev]() { return std::make_unique<Event>(dev); });
+      return cache_[::cms::alpakatools::getDevIndex(dev)].makeOrGet([dev]() { return std::make_unique<Event>(dev); });
     }
 
     // not thread safe, intended to be called only from CUDAService destructor