--- a/src/alpaka/AlpakaCore/StreamCache.h
+++ b/src/alpaka/AlpakaCore/StreamCache.h
@@ -23,7 +23,7 @@ namespace cms::alpakatools {
     // will be returned to the cache by the shared_ptr destructor.
     // This function is thread safe
     ALPAKA_FN_HOST std::shared_ptr<Queue> get(Device const& dev) {
-      return cache_[cms::alpakatools::getDevIndex(dev)].makeOrGet([dev]() { return std::make_unique<Queue>(dev); });
+      return cache_[::cms::alpakatools::getDevIndex(dev)].makeOrGet([dev]() { return std::make_unique<Queue>(dev); });
     }
 
   private: