--- a/src/alpaka/AlpakaDataFormats/BeamSpotAlpaka.h
+++ b/src/alpaka/AlpakaDataFormats/BeamSpotAlpaka.h
@@ -13,8 +13,8 @@ namespace ALPAKA_ACCELERATOR_NAMESPACE {
     BeamSpotAlpaka() = default;
 
     BeamSpotAlpaka(BeamSpotPOD const* data, Queue& queue)
-        : data_d{cms::alpakatools::allocDeviceBuf<BeamSpotPOD>(alpaka::getDev(queue), 1u)} {
-      auto data_h{cms::alpakatools::createHostView<const BeamSpotPOD>(data, 1u)};
+        : data_d{::cms::alpakatools::allocDeviceBuf<BeamSpotPOD>(alpaka::getDev(queue), 1u)} {
+      auto data_h{::cms::alpakatools::createHostView<const BeamSpotPOD>(data, 1u)};
 
       alpaka::memcpy(queue, data_d, data_h, 1u);
       // alpaka::wait(queue);