--- a/src/alpaka/AlpakaDataFormats/TrackingRecHit2DAlpaka.h
+++ b/src/alpaka/AlpakaDataFormats/TrackingRecHit2DAlpaka.h
@@ -1,8 +1,10 @@
 #ifndef CUDADataFormats_TrackingRecHit_interface_TrackingRecHit2DHeterogeneous_h
 #define CUDADataFormats_TrackingRecHit_interface_TrackingRecHit2DHeterogeneous_h
 
-#include "AlpakaDataFormats/TrackingRecHit2DSOAView.h"
+#include <alpaka/alpaka.hpp>
+
 #include "AlpakaCore/alpakaCommon.h"
+#include "AlpakaDataFormats/TrackingRecHit2DSOAView.h"
 
 namespace ALPAKA_ACCELERATOR_NAMESPACE {
 
@@ -20,25 +22,25 @@ namespace ALPAKA_ACCELERATOR_NAMESPACE {
           // NON-OWNING DEVICE POINTERS:
           m_hitsModuleStart(hitsModuleStart),
           // OWNING DEVICE POINTERS:
-          m_xl{cms::alpakatools::allocDeviceBuf<float>(alpaka::getDev(queue), nHits)},
-          m_yl{cms::alpakatools::allocDeviceBuf<float>(alpaka::getDev(queue), nHits)},
-          m_xerr{cms::alpakatools::allocDeviceBuf<float>(alpaka::getDev(queue), nHits)},
-          m_yerr{cms::alpakatools::allocDeviceBuf<float>(alpaka::getDev(queue), nHits)},
-          m_xg{cms::alpakatools::allocDeviceBuf<float>(alpaka::getDev(queue), nHits)},
-          m_yg{cms::alpakatools::allocDeviceBuf<float>(alpaka::getDev(queue), nHits)},
-          m_zg{cms::alpakatools::allocDeviceBuf<float>(alpaka::getDev(queue), nHits)},
-          m_rg{cms::alpakatools::allocDeviceBuf<float>(alpaka::getDev(queue), nHits)},
-          m_iphi{cms::alpakatools::allocDeviceBuf<int16_t>(alpaka::getDev(queue), nHits)},
-          m_charge{cms::alpakatools::allocDeviceBuf<int32_t>(alpaka::getDev(queue), nHits)},
-          m_xsize{cms::alpakatools::allocDeviceBuf<int16_t>(alpaka::getDev(queue), nHits)},
-          m_ysize{cms::alpakatools::allocDeviceBuf<int16_t>(alpaka::getDev(queue), nHits)},
-          m_detInd{cms::alpakatools::allocDeviceBuf<uint16_t>(alpaka::getDev(queue), nHits)},
+          m_xl{::cms::alpakatools::allocDeviceBuf<float>(alpaka::getDev(queue), nHits)},
+          m_yl{::cms::alpakatools::allocDeviceBuf<float>(alpaka::getDev(queue), nHits)},
+          m_xerr{::cms::alpakatools::allocDeviceBuf<float>(alpaka::getDev(queue), nHits)},
+          m_yerr{::cms::alpakatools::allocDeviceBuf<float>(alpaka::getDev(queue), nHits)},
+          m_xg{::cms::alpakatools::allocDeviceBuf<float>(alpaka::getDev(queue), nHits)},
+          m_yg{::cms::alpakatools::allocDeviceBuf<float>(alpaka::getDev(queue), nHits)},
+          m_zg{::cms::alpakatools::allocDeviceBuf<float>(alpaka::getDev(queue), nHits)},
+          m_rg{::cms::alpakatools::allocDeviceBuf<float>(alpaka::getDev(queue), nHits)},
+          m_iphi{::cms::alpakatools::allocDeviceBuf<int16_t>(alpaka::getDev(queue), nHits)},
+          m_charge{::cms::alpakatools::allocDeviceBuf<int32_t>(alpaka::getDev(queue), nHits)},
+          m_xsize{::cms::alpakatools::allocDeviceBuf<int16_t>(alpaka::getDev(queue), nHits)},
+          m_ysize{::cms::alpakatools::allocDeviceBuf<int16_t>(alpaka::getDev(queue), nHits)},
+          m_detInd{::cms::alpakatools::allocDeviceBuf<uint16_t>(alpaka::getDev(queue), nHits)},
           m_averageGeometry{
-              cms::alpakatools::allocDeviceBuf<TrackingRecHit2DSOAView::AverageGeometry>(alpaka::getDev(queue), 1u)},
-          m_hitsLayerStart{cms::alpakatools::allocDeviceBuf<uint32_t>(alpaka::getDev(queue), nHits)},
-          m_hist{cms::alpakatools::allocDeviceBuf<Hist>(alpaka::getDev(queue), 1u)},
+              ::cms::alpakatools::allocDeviceBuf<TrackingRecHit2DSOAView::AverageGeometry>(alpaka::getDev(queue), 1u)},
+          m_hitsLayerStart{::cms::alpakatools::allocDeviceBuf<uint32_t>(alpaka::getDev(queue), nHits)},
+          m_hist{::cms::alpakatools::allocDeviceBuf<Hist>(alpaka::getDev(queue), 1u)},
           // SoA view:
-          m_view{cms::alpakatools::allocDeviceBuf<TrackingRecHit2DSOAView>(alpaka::getDev(queue), 1u)},
+          m_view{::cms::alpakatools::allocDeviceBuf<TrackingRecHit2DSOAView>(alpaka::getDev(queue), 1u)},
           m_view_h{::cms::alpakatools::allocHostBuf<TrackingRecHit2DSOAView>(1u)} {
       // the hits are actually accessed in order only in building
       // if ordering is relevant they may have to be stored phi-ordered by layer or so
@@ -150,8 +152,8 @@ namespace ALPAKA_ACCELERATOR_NAMESPACE {
     }
 #ifdef TODO
     // only the local coord and detector index
-    cms::cuda::host::unique_ptr<uint16_t[]> detIndexToHostAsync(cudaStream_t stream) const;
-    cms::cuda::host::unique_ptr<uint32_t[]> hitsModuleStartToHostAsync(cudaStream_t stream) const;
+    ::cms::alpakatools::host::unique_ptr<uint16_t[]> detIndexToHostAsync(cudaStream_t stream) const;
+    ::cms::alpakatools::host::unique_ptr<uint32_t[]> hitsModuleStartToHostAsync(cudaStream_t stream) const;
 #endif
     auto const* xl() const { return alpaka::getPtrNative(m_xl); }
     auto const* yl() const { return alpaka::getPtrNative(m_yl); }