--- a/src/alpaka/Makefile.deps
+++ b/src/alpaka/Makefile.deps
@@ -4,6 +4,7 @@ alpaka_EXTERNAL_DEPENDS += CUDA
 endif
 AlpakaCore_DEPENDS := Framework
 BeamSpotProducer_DEPENDS := Framework AlpakaCore AlpakaDataFormats DataFormats
+PixelTrackFitting_DEPENDS := Framework AlpakaCore AlpakaDataFormats
 PixelTriplets_DEPENDS := Framework AlpakaCore AlpakaDataFormats
 PixelVertexFinding_DEPENDS := Framework AlpakaCore AlpakaDataFormats DataFormats CondFormats
 SiPixelClusterizer_DEPENDS := Framework AlpakaCore AlpakaDataFormats DataFormats CondFormats