--- a/src/alpaka/plugin-BeamSpotProducer/alpaka/BeamSpotToAlpaka.cc
+++ b/src/alpaka/plugin-BeamSpotProducer/alpaka/BeamSpotToAlpaka.cc
@@ -1,10 +1,10 @@
+#include "AlpakaCore/Product.h"
+#include "AlpakaCore/ScopedContext.h"
 #include "AlpakaDataFormats/BeamSpotAlpaka.h"
-
 #include "Framework/EDProducer.h"
 #include "Framework/Event.h"
 #include "Framework/EventSetup.h"
 #include "Framework/PluginFactory.h"
-#include "AlpakaCore/ScopedContext.h"
 
 namespace ALPAKA_ACCELERATOR_NAMESPACE {
 
@@ -16,18 +16,25 @@ namespace ALPAKA_ACCELERATOR_NAMESPACE {
     void produce(edm::Event& iEvent, const edm::EventSetup& iSetup) override;
 
   private:
-    edm::EDPutTokenT<BeamSpotAlpaka> bsPutToken_;
-    // TO DO: Add implementation of cms::alpaka::Product?
-    // const edm::EDPutTokenT<::cms::alpaka::Product<BeamSpotAlpaka>> bsPutToken_;
+    const edm::EDPutTokenT<::cms::alpakatools::Product<Queue, BeamSpotAlpaka>> bsPutToken_;
+
+    AlpakaHostBuf<BeamSpotPOD> bsHost_;
   };
 
-  BeamSpotToAlpaka::BeamSpotToAlpaka(edm::ProductRegistry& reg) : bsPutToken_{reg.produces<BeamSpotAlpaka>()} {}
+  BeamSpotToAlpaka::BeamSpotToAlpaka(edm::ProductRegistry& reg)
+      : bsPutToken_{reg.produces<::cms::alpakatools::Product<Queue, BeamSpotAlpaka>>()},
+        bsHost_{::cms::alpakatools::allocHostBuf<BeamSpotPOD>(1u)} {
+    alpaka::prepareForAsyncCopy(bsHost_);
+  }
 
   void BeamSpotToAlpaka::produce(edm::Event& iEvent, const edm::EventSetup& iSetup) {
-    auto const& bsRaw = iSetup.get<BeamSpotPOD>();
+    *alpaka::getPtrNative(bsHost_) = iSetup.get<BeamSpotPOD>();
 
     ::cms::alpakatools::ScopedContextProduce<Queue> ctx{iEvent.streamID()};
-    BeamSpotAlpaka bsDevice(&bsRaw, ctx.stream());
+
+    BeamSpotAlpaka bsDevice(ctx.stream());
+    alpaka::memcpy(ctx.stream(), bsDevice.buf(), bsHost_, 1u);
+
     ctx.emplace(iEvent, bsPutToken_, std::move(bsDevice));
   }