--- a/src/alpaka/plugin-PixelTriplets/alpaka/BrokenLineFitOnGPU.cc
+++ b/src/alpaka/plugin-PixelTriplets/alpaka/BrokenLineFitOnGPU.cc
@@ -18,13 +18,13 @@ namespace ALPAKA_ACCELERATOR_NAMESPACE {
         Vec1D::all(numberOfBlocks / 4), Vec1D::all(blockSize));
 
     //  Fit internals
-    auto hitsGPU_ = cms::alpakatools::allocDeviceBuf<double>(
+    auto hitsGPU_ = ::cms::alpakatools::allocDeviceBuf<double>(
         alpaka::getDev(queue), maxNumberOfConcurrentFits_ * sizeof(Rfit::Matrix3xNd<4>) / sizeof(double));
 
-    auto hits_geGPU_ = cms::alpakatools::allocDeviceBuf<float>(
+    auto hits_geGPU_ = ::cms::alpakatools::allocDeviceBuf<float>(
         alpaka::getDev(queue), maxNumberOfConcurrentFits_ * sizeof(Rfit::Matrix6x4f) / sizeof(float));
 
-    auto fast_fit_resultsGPU_ = cms::alpakatools::allocDeviceBuf<double>(
+    auto fast_fit_resultsGPU_ = ::cms::alpakatools::allocDeviceBuf<double>(
         alpaka::getDev(queue), maxNumberOfConcurrentFits_ * sizeof(Rfit::Vector4d) / sizeof(double));
 
     for (uint32_t offset = 0; offset < maxNumberOfTuples; offset += maxNumberOfConcurrentFits_) {