… src/alpaka/AlpakaCore/eventWorkHasCompleted.h
… src/alpaka/plugin-PixelTriplets/alpaka/CAHitNtupletAlpaka.cc
… src/alpaka/plugin-PixelTriplets/alpaka/CAHitNtupletGeneratorKernels.cc
… src/alpaka/plugin-PixelTriplets/alpaka/CAHitNtupletGeneratorKernels.h
… src/alpaka/plugin-PixelTriplets/alpaka/CAHitNtupletGeneratorOnGPU.cc
… src/alpaka/plugin-PixelTriplets/alpaka/CAHitNtupletGeneratorOnGPU.h
… src/alpaka/plugin-PixelTriplets/alpaka/RiemannFitOnGPU.cc
… src/alpaka/plugin-PixelVertexFinding/alpaka/PixelVertexProducerAlpaka.cc
… src/alpaka/plugin-PixelVertexFinding/alpaka/PixelVertexSoAFromAlpaka.cc
… src/alpaka/plugin-PixelVertexFinding/alpaka/gpuVertexFinder.cc
… src/alpaka/plugin-SiPixelClusterizer/alpaka/SiPixelFedCablingMapESProducer.cc
… src/alpaka/plugin-SiPixelClusterizer/alpaka/SiPixelGainCalibrationForHLTESProducer.cc
… src/alpaka/plugin-SiPixelClusterizer/alpaka/SiPixelRawToCluster.cc
… src/alpaka/plugin-SiPixelClusterizer/alpaka/SiPixelRawToClusterGPUKernel.cc
… src/alpaka/plugin-SiPixelRecHits/alpaka/PixelCPEFastESProducer.cc
… src/alpaka/plugin-SiPixelRecHits/alpaka/SiPixelRecHitAlpaka.cc
… src/alpaka/plugin-Validation/alpaka/CountValidator.cc
… src/alpaka/plugin-Validation/alpaka/HistoValidator.cc
… src/alpaka/plugins.txt