@fwyzard 2021-11-01
@makortel this is still in progress - I want to fix all EDProducers before merging it - but it could already use some feedback on the `Product<T>` and `ScopedContext` part.