@makortel 2021-11-01
Changes so far look good to me. I'm only confused a little bit with

> * split `acquire()` and `produce()` in `CAHitNtupletAlpaka`;

Did you mean `PixelTrackSoAFromAlpaka`? (I did not see such changes in `CAHitNtupletAlpaka`)