@fwyzard 2021-11-01
> Changes so far look good to me. 

Thank you for checking !

> I'm only confused a little bit with
> 
> > * split `acquire()` and `produce()` in `CAHitNtupletAlpaka`;
> 
> Did you mean `PixelTrackSoAFromAlpaka`? (I did not see such changes in `CAHitNtupletAlpaka`)

Yes, indeed, I copy-pasted the wrong module's name... I've fixed the commit message and the PR description.