@fwyzard 2021-11-03
Finally working for all backends !

### CUDA
```
fwyzard@gputest-milan-01.cms:/data/user/fwyzard/pixeltrack-standalone$ echo; CUDA_VISIBLE_DEVICES=0 numactl -N 0 ./alpaka --cuda --numberOfThreads 16 --numberOfStreams 16 --validation --maxEvents 1000; echo; for N in 1 2 3 4; do CUDA_VISIBLE_DEVICES=0 numactl -N 0 ./alpaka --cuda --numberOfThreads 16 --numberOfStreams 16 --maxEvents 10000; done

Processing 1000 events, of which 16 concurrently, with 16 threads.
CountValidator: all 1000 events passed validation
 Average relative track difference 0.000897219 (all within tolerance)
 Average absolute vertex difference 0.003 (all within tolerance)
Processed 1000 events in 4.608849e+00 seconds, throughput 216.974 events/s.

Processing 10000 events, of which 16 concurrently, with 16 threads.
Processed 10000 events in 4.151512e+01 seconds, throughput 240.876 events/s.
Processed 10000 events in 4.117855e+01 seconds, throughput 242.845 events/s.
Processed 10000 events in 4.156071e+01 seconds, throughput 240.612 events/s.
Processed 10000 events in 4.168261e+01 seconds, throughput 239.908 events/s.
```

### CPU serial, with a blocking queue
```
fwyzard@gputest-milan-01.cms:/data/user/fwyzard/pixeltrack-standalone$ echo; CUDA_VISIBLE_DEVICES=0 numactl -N 0 ./alpaka --serial --numberOfThreads 16 --numberOfStreams 16 --validation --maxEvents 1000; echo; for N in 1 2 3 4; do CUDA_VISIBLE_DEVICES=0 numactl -N 0 ./alpaka --serial --numberOfThreads 16 --numberOfStreams 16 --maxEvents 10000; done

Processing 1000 events, of which 16 concurrently, with 16 threads.
CountValidator: all 1000 events passed validation
 Average relative track difference 0.00133904 (all within tolerance)
 Average absolute vertex difference 0.005 (all within tolerance)
Processed 1000 events in 1.786370e+00 seconds, throughput 559.794 events/s.

Processing 10000 events, of which 16 concurrently, with 16 threads.
Processed 10000 events in 1.562877e+01 seconds, throughput 639.845 events/s.
Processed 10000 events in 1.583063e+01 seconds, throughput 631.687 events/s.
Processed 10000 events in 1.565146e+01 seconds, throughput 638.918 events/s.
Processed 10000 events in 1.601298e+01 seconds, throughput 624.493 events/s.
```

### CPU parallel blocks with TBB, with a non-blocking queue
```
fwyzard@gputest-milan-01.cms:/data/user/fwyzard/pixeltrack-standalone$ echo; CUDA_VISIBLE_DEVICES=0 numactl -N 0 ./alpaka --tbb --numberOfThreads 16 --numberOfStreams 16 --validation --maxEvents 1000; echo; for N in 1 2 3 4; do CUDA_VISIBLE_DEVICES=0 numactl -N 0 ./alpaka --tbb --numberOfThreads 16 --numberOfStreams 16 --maxEvents 10000; done

Processing 1000 events, of which 16 concurrently, with 16 threads.
CountValidator: all 1000 events passed validation
 Average relative track difference 0.00163275 (all within tolerance)
 Average absolute vertex difference 0.002 (all within tolerance)
Processed 1000 events in 7.997295e+00 seconds, throughput 125.042 events/s.

Processing 10000 events, of which 16 concurrently, with 16 threads.
Processed 10000 events in 7.524938e+01 seconds, throughput 132.891 events/s.
Processed 10000 events in 7.617038e+01 seconds, throughput 131.285 events/s.
Processed 10000 events in 7.553579e+01 seconds, throughput 132.388 events/s.
Processed 10000 events in 7.450857e+01 seconds, throughput 134.213 events/s.
```