@fwyzard 2021-11-04
Reimplement the relevant modules as "external work" producers, and split their work into the `acquire()` and `produce()` methods:
  - SiPixelRawToCluster
  - PixelTrackSoAFromAlpaka
  - PixelVertexSoAFromAlpaka
  - HistoValidator

Fix the ScopedContext class hierarchy:
  - emplace cms::alpakatools::Product<Queue, T> instead of T;
  - fix the callback to release the WaitingTaskHolder when the asynchronous work is done.

Make all modules consume and produce asynchronous products via cms::alpakatools::Product<Queue, T>.

Remove the unnecessary queue synchronisations, and document why the others are still needed.

Code clan up:
  - move PixelTrackSoAFromAlpaka under plugin-PixelTrackFitting/;
  - replace eventWorkHasCompleted with alpaka::isComplete;
  - fix the use of the ::cms::alpakatools namespace;
  - other minor clean up.