@fwyzard 2021-11-04
# Impact on the performance

## before

### `alpaka --cuda`
```bash
$ CUDA_VISIBLE_DEVICES=0 numactl -N 0 ./alpaka --cuda --numberOfThreads 16 --numberOfStreams 16 --validation --maxEvents 1000

Processing 1000 events, of which 16 concurrently, with 16 threads.
CountValidator: all 1000 events passed validation
 Average relative track difference 0.000885944 (all within tolerance)
 Average absolute vertex difference 0.003 (all within tolerance)
Processed 1000 events in 4.685915e+00 seconds, throughput 213.405 events/s.
```

```bash
$ for N in {1..4}; do CUDA_VISIBLE_DEVICES=0 numactl -N 0 ./alpaka --cuda --numberOfThreads 16 --numberOfStreams 16 --maxEvents 10000; done

Processing 10000 events, of which 16 concurrently, with 16 threads.
Processed 10000 events in 4.067341e+01 seconds, throughput 245.861 events/s.
Processed 10000 events in 4.083298e+01 seconds, throughput 244.9 events/s.
Processed 10000 events in 4.104474e+01 seconds, throughput 243.637 events/s.
Processed 10000 events in 4.101111e+01 seconds, throughput 243.836 events/s.
```

### `alpaka --serial`
```bash
$ numactl -N 0 ./alpaka --serial --numberOfThreads 16 --numberOfStreams 16 --validation --maxEvents 1000

Processing 1000 events, of which 16 concurrently, with 16 threads.
CountValidator: all 1000 events passed validation
 Average relative track difference 0.00133904 (all within tolerance)
 Average absolute vertex difference 0.005 (all within tolerance)
Processed 1000 events in 1.859995e+00 seconds, throughput 537.636 events/s.
```

```bash
$ for N in {1..4}; do numactl -N 0 ./alpaka --serial --numberOfThreads 16 --numberOfStreams 16 --maxEvents 10000; done

Processing 10000 events, of which 16 concurrently, with 16 threads.
Processed 10000 events in 1.601715e+01 seconds, throughput 624.331 events/s.
Processed 10000 events in 1.590361e+01 seconds, throughput 628.788 events/s.
Processed 10000 events in 1.572095e+01 seconds, throughput 636.094 events/s.
Processed 10000 events in 1.614289e+01 seconds, throughput 619.468 events/s.
```

### `alpaka --tbb` with Felice's `atomic_ref.patch`
```bash
$ numactl -N 0 ./alpaka --tbb --numberOfThreads 16 --numberOfStreams 16 --validation --maxEvents 1000

Processing 1000 events, of which 16 concurrently, with 16 threads.
CountValidator: all 1000 events passed validation
 Average relative track difference 0.00134253 (all within tolerance)
 Average absolute vertex difference 0.005 (all within tolerance)
Processed 1000 events in 2.367957e+00 seconds, throughput 422.305 events/s.
```

```bash
$ for N in {1..4} do numactl -N 0 ./alpaka --tbb --numberOfThreads 16 --numberOfStreams 16 --maxEvents 1000; done

Processing 1000 events, of which 16 concurrently, with 16 threads.
Processed 1000 events in 2.228229e+00 seconds, throughput 448.787 events/s.
Processed 1000 events in 2.243764e+00 seconds, throughput 445.68 events/s.
Processed 1000 events in 2.215630e+00 seconds, throughput 451.339 events/s.
Processed 1000 events in 2.247022e+00 seconds, throughput 445.033 events/s.
```

## after

### `alpaka --cuda`
```bash
$ CUDA_VISIBLE_DEVICES=0 numactl -N 0 ./alpaka --cuda --numberOfThreads 16 --numberOfStreams 16 --validation --maxEvents 1000

Processing 1000 events, of which 16 concurrently, with 16 threads.
CountValidator: all 1000 events passed validation
 Average relative track difference 0.000912222 (all within tolerance)
 Average absolute vertex difference 0.001 (all within tolerance)
Processed 1000 events in 4.596381e+00 seconds, throughput 217.562 events/s.
```

```bash
$ for N in {1..4}; do CUDA_VISIBLE_DEVICES=0 numactl -N 0 ./alpaka --cuda --numberOfThreads 16 --numberOfStreams 16 --maxEvents 10000; done

Processing 10000 events, of which 16 concurrently, with 16 threads.
Processed 10000 events in 4.151018e+01 seconds, throughput 240.905 events/s.
Processed 10000 events in 4.105583e+01 seconds, throughput 243.571 events/s.
Processed 10000 events in 4.081506e+01 seconds, throughput 245.008 events/s.
Processed 10000 events in 4.088302e+01 seconds, throughput 244.6 events/s.
```

### `alpaka --serial`
```bash
$ numactl -N 0 ./alpaka --serial --numberOfThreads 16 --numberOfStreams 16 --validation --maxEvents 1000

Processing 1000 events, of which 16 concurrently, with 16 threads.
CountValidator: all 1000 events passed validation
 Average relative track difference 0.00133904 (all within tolerance)
 Average absolute vertex difference 0.005 (all within tolerance)
Processed 1000 events in 1.822133e+00 seconds, throughput 548.807 events/s.
```

```bash
$ for N in {1..4}; do numactl -N 0 ./alpaka --serial --numberOfThreads 16 --numberOfStreams 16 --maxEvents 10000; done

Processing 10000 events, of which 16 concurrently, with 16 threads.
Processed 10000 events in 1.607611e+01 seconds, throughput 622.041 events/s.
Processed 10000 events in 1.609473e+01 seconds, throughput 621.321 events/s.
Processed 10000 events in 1.614500e+01 seconds, throughput 619.387 events/s.
Processed 10000 events in 1.574037e+01 seconds, throughput 635.309 events/s.
```


### `alpaka --tbb` with Felice's `atomic_ref.patch`

```bash
$ numactl -N 0 ./alpaka --tbb --numberOfThreads 16 --numberOfStreams 16 --validation --maxEvents 1000

Processing 1000 events, of which 16 concurrently, with 16 threads.
CountValidator: all 1000 events passed validation
 Average relative track difference 0.00157396 (all within tolerance)
 Average absolute vertex difference 0.006 (all within tolerance)
Processed 1000 events in 8.184133e+00 seconds, throughput 122.188 events/s.

```bash
$ for N in {1..4} do numactl -N 0 ./alpaka --tbb --numberOfThreads 16 --numberOfStreams 16 --maxEvents 1000; done

Processing 1000 events, of which 16 concurrently, with 16 threads.
Processed 1000 events in 7.905449e+00 seconds, throughput 126.495 events/s.
Processed 1000 events in 7.780583e+00 seconds, throughput 128.525 events/s.
Processed 1000 events in 7.907862e+00 seconds, throughput 126.456 events/s.
Processed 1000 events in 7.783735e+00 seconds, throughput 128.473 events/s.
```

The performance of the TBB backend can be recovered and improved
  - with the changes to the atomic hierachy by Felice and Angela
  - by using a blocking queue